+++ example/main.dart
@@ -1,30 +1,3 @@
-
-A pure Dart implementation of the Firebase admin sdk
-
-Currently, only supports admin methods for the following firebase services:
-
-* authentication
-* realtime database
-
-## Example using service account
-
-```dart
-import 'package:firebase_admin/firebase_admin.dart';
-
-main() async {
-  var app = FirebaseAdmin.instance.initializeApp(AppOptions(
-    credential: ServiceAccountCredential('service-account.json'),
-  ));
-
-  var link = await app.auth().generateSignInWithEmailLink('jane@doe.com',
-      ActionCodeSettings(url: 'https://example.com'));
-
-  print(link);
-}
-```
-## Example using default credentials
-
-```dart
 import 'package:firebase_admin/firebase_admin.dart';
 import 'package:firebase_admin/src/credential.dart';
 
@@ -53,12 +26,3 @@ void main() async {
     print(e.message);
   }
 }
-
-```
-
-
-## Features and bugs
-
-Please file feature requests and bugs at the [issue tracker][tracker].
-
-[tracker]: https://github.com/appsup-dart/firebase_admin/issues