+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+## 0.1.1
+
+- look for default credentials in Firebase CLI configurations
+- get credentials by openid login
+
+
 ## 0.1.0
 
 - admin sdk for firebase realtime database